fix: delete unnecessary files and rename mvc folder
