--- src/MVC/model/readme.md
@@ -1,1 +1,0 @@
-#### folder for functions for general logics app
