--- src/MVC/view/readme.md
@@ -1,1 +1,0 @@
-#### folder for functions remder, update view and DOM manipulations
