--- src/assets/fonts/readme.md
@@ -1,2 +1,0 @@
-### IIn this folder we can save fonts and from here we can link these fonts to ../../styles/font.css
-
