--- src/assets/lib/readme.md
@@ -1,1 +1,0 @@
-### In this folder you can save minify libraries like bootstrap, tailwind and so one
