+++ src/components/view/authorization/NewUser.ts
