+++ src/components/view/authorization/_authorization.scss
