+++ src/components/view/footer/_footer.scss
