+++ src/components/view/footer/footer.ts
