+++ src/components/view/navPanel/_navElements.scss
