+++ src/components/view/pages/events/EventsPage.ts
