+++ src/components/view/pages/events/_events.scss
