+++ src/components/view/pages/events/eventsType.ts
