+++ src/components/view/pages/main/MainPage.ts
