+++ src/components/view/pages/main/TStype/mainTypes.ts
