+++ src/components/view/pages/main/_mainPage.scss
