+++ src/components/view/pages/main/_responsive-main.scss
