+++ src/components/view/pages/notFound/_404.scss
