+++ src/components/view/welcome-popup/_welcome-popup.scss
