+++ src/components/view/welcome-popup/welcome-popup.ts
