+++ src/index.ts
@@ -1,6 +1,6 @@
 import './styles/main.scss';
 
-import App from './mvc/app/app';
+import App from './components/app/app';
 
 const app = new App();
 app.renderBaseElements();
